backend dos novos graficos

+++ api/src/main/kotlin/project/api/app/dashboard/CandidatosPorVagaDTO.kt
@@ -2,5 +2,5 @@ package project.api.app.dashboard
 
 data class CandidatosPorVagaDTO(
     val vaga: String,
-    val totalCandidatos: Long
+    val totalCandidatos: Int
 )
