+++ api/src/main/kotlin/project/api/app/dashboard/TempoMedioPreenchimentoDTO.kt
@@ -2,5 +2,5 @@ package project.api.app.dashboard
 
 data class TempoMedioPreenchimentoDTO(
     val mes: Int,
-    val dias: Double
+    val dias: Int
 )
