+++ api/src/main/kotlin/project/api/app/dashboard/TipoContratoDTO.kt
@@ -2,5 +2,5 @@ package project.api.app.dashboard
 
 data class TipoContratoDTO(
     val contrato: String,
-    val total: Long
+    val total: Int
 )
