+++ api/src/main/kotlin/project/api/app/dashboard/VagasPorMesDTO.kt
@@ -2,5 +2,5 @@ package project.api.app.dashboard
 
 data class VagasPorMesDTO(
     val mes: Int,
-    val quantidade: Long
+    val quantidade: Int
 )
